+++ Source/PangleAdapterConfiguration.swift
@@ -18,7 +18,7 @@ import PAGAdSDK
     /// last digit is the adapter's build version, and intermediate digits are the partner SDK's version.
     /// Format: `<Chartboost Mediation major version>.<Partner major version>.<Partner minor version>.<Partner patch version>.
     /// <Partner build version>.<Adapter build version>` where `.<Partner build version>` is optional.
-    @objc public static let adapterVersion = "5.7.8.5.0.0"
+    @objc public static let adapterVersion = "5.7.9.0.0.0"
 
     /// The partner's unique identifier.
     @objc public static let partnerID = "pangle"
@@ -28,22 +28,11 @@ import PAGAdSDK
 
     /// Use to manually set the consent status on the Pangle SDK.
     /// This is generally unnecessary as the Mediation SDK will set the consent status automatically based on the latest consent info.
-    @objc public static func setGDPRConsentOverride(_ consent: PAGGDPRConsentType) {
-        isGDPRConsentOverridden = true
-        PAGConfig.share().gdprConsent = consent
-        log("GDPR consent override set to \(consent)")
-    }
-
-    /// Use to manually set the consent status on the Pangle SDK.
-    /// This is generally unnecessary as the Mediation SDK will set the consent status automatically based on the latest consent info.
     @objc public static func setPAConsentTypeOverride(_ consentType: PAGPAConsentType) {
         isPAConsentOverridden = true
         PAGConfig.share().paConsent = consentType
         log("PA consent override set to \(consentType)")
     }
-
-    /// Internal flag that indicates if the GDPR consent has been overridden by the publisher.
-    private(set) static var isGDPRConsentOverridden = false
 
     /// Internal flag that indicates if the PA consent has been overridden by the publisher.
     private(set) static var isPAConsentOverridden = false